[FIX] #8 실내정원용 식물 목록 구조 변경
